+++ lib/fog/azurerm/models/compute/disk_create_option.rb
@@ -1,0 +1,16 @@
+module Fog
+  module ARM
+    module Compute
+      module Models
+        module DiskCreateOption
+          FromImage = 'fromImage'.freeze
+          Empty = 'empty'.freeze
+          Attach = 'attach'.freeze
+          Import = 'Import'.freeze
+          Copy = 'Copy'.freeze
+          Restore = 'Restore'.freeze
+        end
+      end
+    end
+  end
+end