+++ test/api_stub.rb
@@ -5,6 +5,7 @@ module ApiStub
       autoload :Server, File.expand_path('api_stub/models/compute/server', __dir__)
       autoload :AvailabilitySet, File.expand_path('api_stub/models/compute/availability_set', __dir__)
       autoload :VirtualMachineExtension, File.expand_path('api_stub/models/compute/virtual_machine_extension', __dir__)
+      autoload :ManagedDisk, File.expand_path('api_stub/models/compute/managed_disk', __dir__)
     end
 
     module Resources
@@ -67,6 +68,7 @@ module ApiStub
       autoload :AvailabilitySet, File.expand_path('api_stub/requests/compute/availability_set', __dir__)
       autoload :VirtualMachine, File.expand_path('api_stub/requests/compute/virtual_machine', __dir__)
       autoload :VirtualMachineExtension, File.expand_path('api_stub/requests/compute/virtual_machine_extension', __dir__)
+      autoload :ManagedDisk, File.expand_path('api_stub/requests/compute/managed_disk', __dir__)
     end
 
     module Resources